fix(windows): make workflow revision publishing portable
