+++ services/tuttid/data/workspace/workflow_revision_directory_sync_windows.go
@@ -1,0 +1,9 @@
+//go:build windows
+
+package workspace
+
+// Windows does not provide portable directory fsync through os.File. The
+// revision file itself is synced before publication.
+func syncWorkflowRevisionDirectory(string) error {
+	return nil
+}
